Rename file with error types
